requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Flask==2.3.0
python-dotenv==1.0.0
Werkzeug==2.3.0
Jinja2==3.1.2
click==8.1.3
itsdangerous==2.1.2
